+++ sync-for-payables/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v0.2.1] sync-for-payables
 ### Releases
 - [PyPI v0.2.1] https://pypi.org/project/codat-sync-for-payables/0.2.1 - sync-for-payables
+
+## 2023-09-13 00:17:24
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payables.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.2] sync-for-payables
+### Releases
+- [PyPI v0.2.2] https://pypi.org/project/codat-sync-for-payables/0.2.2 - sync-for-payables